+++ client/cache/client.go
@@ -19,21 +19,3 @@ type RedisClient interface {
 	RPush(ctx context.Context, key string, values []interface{}) error
 	LRange(ctx context.Context, key string, start, stop int) ([]interface{}, error)
 }
-
-// KeyComparator comparator function type
-type KeyComparator func(a, b string) bool
-
-// ScanOptions scan options
-type ScanOptions struct {
-	KeyComparator *KeyComparator
-}
-
-// ScanOption scan option
-type ScanOption func(*ScanOptions)
-
-// WithKeyComparator key comparator function
-func WithKeyComparator(keyComparator KeyComparator) ScanOption {
-	return func(options *ScanOptions) {
-		options.KeyComparator = &keyComparator
-	}
-}
